chore(M2Mqtt): Removes obsolete .net support, updates to 4.6.1 and .net core
I didn't see the point of supporting obsolete .net versions

Also added a target, so the library can be hands-on tested and unit
tested easily.

So far the target only subscribes to all topics and spits them out to
the console, but it wil be expanded with arguments to run a series of
tests, or run as a one-shot publisher, or subscribe to a topic and spit
that out to the console.